+++ Changelog.md
@@ -3,6 +3,7 @@ Lilu Changelog
 
 #### v1.2.7
 - Added support for detecting optimus switch-off
+- Added Sanitize target with ubsan support (thx to NetBSD)
 - Fixed multiple Mach-O parsing issues
 - Fixed support of PCI devices without compatible property